requirements.txt
Django==4.2.10
python-dotenv==1.0.0
opencv-python-headless==4.8.0.76
pytesseract==0.3.10
PyPDF2==3.0.1
Pillow==9.5.0
numpy==1.24.3
pytest==7.4.0
pytest-django==4.5.2
black==23.7.0
flake8==6.1.0
isort==5.12.0
openai>=1.0.0